backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.6.0

# Google Calendar API
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# Date/Time parsing
python-dateutil==2.9.0
pytz==2024.2
dateparser==1.2.0

# HTTP client
httpx==0.27.0
